modified update and delete controllers and queries relates #18

+++ server/database/queries/products/deleteProductQuery.ts
@@ -1,8 +1,9 @@
 import { Product } from '../../../models';
 
-const deleteProductQuery = async (id:string) => Product.destroy({
+const deleteProductQuery = async (id: string, user_id: number) => Product.destroy({
   where: {
     id,
+    user_id,
     deletedAt: null,
   },
 });
